Refactor: ClubApplicationStatus를 범용 ApplicationStatus 로 변경 및 패키지 이동
