+++ src/main/java/com/greedy/mokkoji/api/clubapplication/dto/response/ClubApplicationResponse.java
@@ -1,7 +1,7 @@
 package com.greedy.mokkoji.api.clubapplication.dto.response;
 
 import com.greedy.mokkoji.db.clubapplication.entity.ClubApplication;
-import com.greedy.mokkoji.enums.clubApplication.ClubApplicationStatus;
+import com.greedy.mokkoji.enums.application.ApplicationStatus;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.time.LocalDateTime;
@@ -10,7 +10,7 @@ public record ClubApplicationResponse(
         @Schema(example = "1") Long clubApplicationId,
         @Schema(example = "세종대학교") String universityName,
         @Schema(example = "그리디") String clubName,
-        @Schema(example = "PENDING") ClubApplicationStatus status,
+        @Schema(example = "PENDING") ApplicationStatus status,
         @Schema(example = "서류 미비") String rejectReason,
         @Schema(example = "2026-05-18T13:00:00") LocalDateTime createdAt
 ) {
