+++ src/main/java/com/greedy/mokkoji/db/clubapplication/repository/ClubApplicationRepository.java
@@ -3,7 +3,7 @@ package com.greedy.mokkoji.db.clubapplication.repository;
 import com.greedy.mokkoji.db.clubapplication.entity.ClubApplication;
 import com.greedy.mokkoji.db.university.entity.University;
 import com.greedy.mokkoji.db.user.entity.User;
-import com.greedy.mokkoji.enums.clubApplication.ClubApplicationStatus;
+import com.greedy.mokkoji.enums.application.ApplicationStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -12,7 +12,7 @@ import java.util.List;
 @Repository
 public interface ClubApplicationRepository extends JpaRepository<ClubApplication, Long>, ClubApplicationRepositoryCustom {
 
-    boolean existsByApplicantAndUniversityAndStatusNot(User applicant, University university, ClubApplicationStatus status);
+    boolean existsByApplicantAndUniversityAndStatusNot(User applicant, University university, ApplicationStatus status);
 
     List<ClubApplication> findByApplicant(User applicant);
 }
